+++ crates/jsonrtl-api/src/main.rs
@@ -4,5 +4,5 @@ use std::net::{Ipv4Addr, SocketAddr};
 async fn main() -> std::io::Result<()> {
     let address = SocketAddr::from((Ipv4Addr::LOCALHOST, 3000));
     let listener = tokio::net::TcpListener::bind(address).await?;
-    axum::serve(listener, logic_kernel_api::RouterBuilder::new().finish()).await
+    axum::serve(listener, jsonrtl_api::RouterBuilder::new().finish()).await
 }